# protobuf/api.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "api.proto",
    "version": "version not set"
  },
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/healthz": {
      "get": {
        "operationId": "Healthz",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/APIStatusMessage"
            }
          }
        },
        "tags": [
          "API"
        ]
      }
    },
    "/v1/queue/{queueID}": {
      "get": {
        "operationId": "LoadTask",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/APITask"
            }
          }
        },
        "parameters": [
          {
            "name": "queueID",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "expirationSec",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "phase",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          }
        ],
        "tags": [
          "API"
        ]
      },
      "post": {
        "operationId": "PushTask",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/APIStatusMessage"
            }
          }
        },
        "parameters": [
          {
            "name": "queueID",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/APITask"
            }
          }
        ],
        "tags": [
          "API"
        ]
      }
    },
    "/v1/queue/{queueID}/progress": {
      "get": {
        "operationId": "Progress",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/APIQueueProgress"
            }
          }
        },
        "parameters": [
          {
            "name": "queueID",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "expirationSec",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "phase",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          }
        ],
        "tags": [
          "API"
        ]
      }
    },
    "/v1/queue/{queueID}/{phase}/{limit}": {
      "get": {
        "operationId": "Peek",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/APITaskList"
            }
          }
        },
        "parameters": [
          {
            "name": "queueID",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "phase",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "limit",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "expirationSec",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          }
        ],
        "tags": [
          "API"
        ]
      }
    },
    "/v1/queue/{queueID}/{taskID}": {
      "put": {
        "operationId": "Complete",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/APIStatusMessage"
            }
          }
        },
        "parameters": [
          {
            "name": "queueID",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "taskID",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "API"
        ]
      }
    }
  },
  "definitions": {
    "APIQueueProgress": {
      "type": "object",
      "properties": {
        "incomingCount": {
          "type": "integer",
          "format": "int32"
        },
        "receivedCount": {
          "type": "integer",
          "format": "int32"
        },
        "runningCount": {
          "type": "integer",
          "format": "int32"
        },
        "expiredCount": {
          "type": "integer",
          "format": "int32"
        },
        "completedCount": {
          "type": "integer",
          "format": "int32"
        },
        "failedCount": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "APIStatus": {
      "type": "string",
      "enum": [
        "SUCCESS",
        "FAILURE",
        "UNKNOWN"
      ],
      "default": "SUCCESS"
    },
    "APIStatusMessage": {
      "type": "object",
      "properties": {
        "status": {
          "$ref": "#/definitions/APIStatus"
        },
        "detail": {
          "type": "string"
        }
      }
    },
    "APITask": {
      "type": "object",
      "properties": {
        "queueID": {
          "type": "string",
          "title": "reserved 1; // we took queueID out\nreserved \"queueID\";"
        },
        "taskID": {
          "type": "string"
        },
        "body": {
          "type": "string"
        },
        "envs": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "expirationSec": {
          "type": "integer",
          "format": "int32"
        }
      },
      "title": "Task is the definition of a task"
    },
    "APITaskList": {
      "type": "object",
      "properties": {
        "totalItems": {
          "type": "integer",
          "format": "int32"
        },
        "tasks": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/APITask"
          }
        }
      }
    }
  }
}